pipeline/requirements.txt
pytest
pylint
python-dotenv
psycopg2-binary
confluent_kafka